+++ CHANGES.rst
@@ -4,6 +4,9 @@ CHANGES
 2.0.2 (unreleased)
 ------------------
 
+- Removed ext_editor and visible_ids preferences.
+  [davisagli]
+
 - More explicit ZCML package includes. At least, the inclusion of
   plone.formwidget.namedfile fixes a problem in Dexterity-less setups, where
   the @@personal-information form couldn't be rendered because NamedBlobImage